fixed sorting deck path bug in includes

+++ src/test_assets.cpp
@@ -1,8 +1,8 @@
-#include "sorting_deck.h"
+#include "../sorting_deck.h"
 
 
 bool equal(int lhs[], int rhs[], int n){
-  for (size_t i = 0; i < n; i++) {
+  for (short i = 0; i < n; i++) {
     if(lhs[i] != rhs[i])
       return false;
   }
